Move slot / bookings checks / fixes to their own file
